table: disable cache for reversed reads
Before we implement efficient reverse reads (which don't fetch entire
partitions into memory, causing out-of-memory errors) in the sstable
layer, we disable cache for them: the first iteration will not support
fast forwading which is required by the cache. Some changes to the cache
itself may also be required for reverse mode.