table: make_reader: push reversing reader down below combined reader
To implement efficient reversed queries, we need to reverse directly at
the level of memtable and sstable readers. In order to work on the two
independently, we split the `make_reversing_reader` invocation above the
combined reader into a set of invocations below the combined reader:
one invocation for each memtable and one invocation for the sstable reader;
the combined reader is now combining readers that are already reversed.