memtable, partition_snapshot_reader: read from partition in reverse
In this commit, I add the ability to read from partition snapshots in
reverse order. Memtable decides based on options in `slice`, whether
read forward or in reverse.

The reversing partition snapshot reader maintains two schemas - one that
is the reversed schema for the output, and the other one (forward one),
which is used to access the memtable tree (which needs to be the same as
the schema used to create memtable).